ensure saml attributes are formatted/encoded
